Split test_pull_request_review_processor into focused modules
CodeScene flagged Low Cohesion (18 responsibilities across 52 functions, threshold = 4). Split the monolithic test module into eight focused files grouped by area of behavior:

- test_pull_request_review_processor_skip.py - early-exit preconditions and no-feedback path
- test_pull_request_review_processor_triage.py - triage execute() dispatch, JSON parsing, prompt/response logging
- test_pull_request_review_processor_fix.py - fix-group dispatch and post-fix state marking
- test_pull_request_review_processor_format.py - _format_all_feedback rendering
- test_pull_request_review_processor_feedback_helpers.py - _get_new_feedback / _get_feedback_by_ids / _determine_comment_type
- test_pull_request_review_processor_reply.py - i2code-marker prefix on push-and-reply and clarification replies
- test_pull_request_review_processor_filtering.py - self-comment and resolved-thread filters
- test_parse_owner_repo.py - parse_owner_repo URL parser

The shared make_processor factory is extracted to pull_request_review_processor_helpers.py so each test file imports only the helpers it needs.

Co-authored by Claude Code
