--- tests/implement/test_pull_request_review_processor.py
@@ -1,749 +1,0 @@
-"""Tests for PullRequestReviewProcessor class."""
-
-import json
-import os
-
-import pytest
-
-from i2code.implement.pull_request_review_processor import PullRequestReviewProcessor, parse_owner_repo
-from i2code.implement.claude_runner import (
-    CapturedOutput,
-    ClaudeCodeCommand,
-    ClaudeResult,
-    _parse_stream_json_output,
-)
-from i2code.implement.command_builder import CommandBuilder
-
-from fake_git_repository import FakeGitRepository
-from fake_github_client import FakeGitHubClient
-from fake_workflow_state import FakeWorkflowState
-from fake_claude_runner import FakeClaudeRunner
-from i2code.implement.implement_opts import ImplementOpts
-
-FIXTURE_DIR = os.path.join(os.path.dirname(__file__), "fixtures")
-PR6_FEEDBACK_FILE = os.path.join(FIXTURE_DIR, "pr6_feedback.json")
-PR6_TRIAGE_STDOUT_FILE = os.path.join(FIXTURE_DIR, "pr6_triage_stdout.txt")
-
-
-def _make_processor(**overrides):
-    """Create a PullRequestReviewProcessor with common test defaults.
-
-    All parameters are keyword-only overrides to the defaults dict.
-    """
-    defaults = dict(
-        pr_number=42, pushed=True, branch="idea/test/01-slice",
-        working_tree_dir="/tmp/worktree", non_interactive=True,
-        mock_claude=None, skip_ci_wait=True, ci_timeout=600,
-        pr_url="https://github.com/org/repo/pull/42",
-        comments=None, reviews=None, conversation_comments=None,
-    )
-    defaults.update(overrides)
-    d = defaults
-
-    fake_gh = FakeGitHubClient()
-    if d["pr_number"] and d["pr_url"]:
-        fake_gh.set_pr_url(d["pr_number"], d["pr_url"])
-    if d["comments"] is not None:
-        fake_gh.set_pr_comments(d["pr_number"], d["comments"])
-    if d["reviews"] is not None:
-        fake_gh.set_pr_reviews(d["pr_number"], d["reviews"])
-    if d["conversation_comments"] is not None:
-        fake_gh.set_pr_conversation_comments(d["pr_number"], d["conversation_comments"])
-
-    fake_repo = FakeGitRepository(working_tree_dir=d["working_tree_dir"], gh_client=fake_gh)
-    fake_repo.pr_number = d["pr_number"]
-    fake_repo.branch = d["branch"]
-    fake_repo.set_pushed(d["pushed"])
-
-    fake_state = FakeWorkflowState()
-    fake_claude = FakeClaudeRunner()
-
-    opts = ImplementOpts(
-        idea_directory="/tmp/idea",
-        non_interactive=d["non_interactive"],
-        mock_claude=d["mock_claude"],
-        skip_ci_wait=d["skip_ci_wait"],
-        ci_timeout=d["ci_timeout"],
-    )
-
-    processor = PullRequestReviewProcessor(
-        opts=opts, git_repo=fake_repo, state=fake_state, claude_runner=fake_claude,
-    )
-
-    return processor, fake_gh, fake_repo, fake_state, fake_claude
-
-
-def _make_skip_processor(pr_number=None, pushed=False):
-    """Create a minimal processor for testing skip conditions."""
-    fake_gh = FakeGitHubClient()
-    fake_repo = FakeGitRepository(working_tree_dir="/tmp", gh_client=fake_gh)
-    if pr_number:
-        fake_repo.pr_number = pr_number
-    fake_repo.set_pushed(pushed)
-    processor = PullRequestReviewProcessor(
-        opts=ImplementOpts(idea_directory="/tmp/idea"),
-        git_repo=fake_repo,
-        state=FakeWorkflowState(),
-        claude_runner=FakeClaudeRunner(),
-    )
-    return processor, fake_gh
-
-
-def _triage_with_fix(comment_ids, description="Fix issue"):
-    return json.dumps({
-        "will_fix": [{"comment_ids": comment_ids, "description": description}],
-        "needs_clarification": [],
-    })
-
-
-def _make_fix_processor(comments, triage_json, non_interactive=True, **extra):
-    """Create a processor configured for fix-group tests with head SHA advancement."""
-    triage_result = ClaudeResult(
-        returncode=0,
-        output=CapturedOutput(triage_json),
-        result_text=triage_json,
-    )
-    fix_result = ClaudeResult(returncode=0)
-
-    processor, _, fake_repo, fake_state, fake_claude = _make_processor(
-        comments=comments, non_interactive=non_interactive, **extra,
-    )
-    fake_repo.set_head_sha("aaa111")
-
-    def advance_head():
-        fake_repo.set_head_sha("bbb222")
-
-    fake_claude.set_results([triage_result, fix_result])
-    fake_claude.set_side_effects([lambda: None, advance_head])
-
-    return processor, fake_repo, fake_state, fake_claude
-
-
-@pytest.mark.unit
-class TestPullRequestReviewProcessorSkipConditions:
-    """PullRequestReviewProcessor skips feedback when preconditions not met."""
-
-    def test_skips_when_no_pr_number(self):
-        processor, fake_gh = _make_skip_processor(pushed=True)
-        assert processor.process_feedback() is False
-        assert not any(c[0] == "get_pr_url" for c in fake_gh.calls)
-
-    def test_skips_when_not_pushed(self):
-        processor, fake_gh = _make_skip_processor(pr_number=42)
-        assert processor.process_feedback() is False
-        assert not any(c[0] == "fetch_pr_comments" for c in fake_gh.calls)
-
-
-@pytest.mark.unit
-class TestProcessPrFeedbackNoFeedback:
-    """process_pr_feedback returns (False, False) when no new feedback."""
-
-    def test_no_new_feedback_returns_false_false(self):
-        processor, _, _, _, _ = _make_processor(comments=[], reviews=[], conversation_comments=[])
-        had_feedback, made_changes = processor.process_pr_feedback()
-        assert had_feedback is False
-        assert made_changes is False
-
-
-@pytest.mark.unit
-class TestProcessPrFeedbackTriage:
-    """process_pr_feedback triages feedback using claude_runner.execute()."""
-
-    def test_triage_uses_execute_with_command_dataclass(self):
-        """Non-mock triage dispatches execute() with the CommandBuilder triage command."""
-        triage_json = json.dumps({"will_fix": [], "needs_clarification": []})
-        triage_result = ClaudeResult(
-            returncode=0,
-            output=CapturedOutput(triage_json),
-            result_text=triage_json,
-        )
-
-        feedback_comment = {"id": 1, "body": "fix this", "user": {"login": "u"}}
-        processor, _, fake_repo, _, fake_claude = _make_processor(
-            comments=[feedback_comment],
-        )
-        fake_claude.set_result(triage_result)
-
-        processor.process_pr_feedback()
-
-        assert len(fake_claude.calls) == 1
-        method, cmd, cwd = fake_claude.calls[0]
-        assert method == "execute"
-        assert isinstance(cmd, ClaudeCodeCommand)
-        feedback_content = PullRequestReviewProcessor._format_all_feedback(
-            [feedback_comment], [], [],
-        )
-        expected = CommandBuilder().build_triage_command(
-            feedback_content,
-            cwd=fake_repo.working_tree_dir,
-            interactive=False,
-        )
-        assert cmd == expected
-        assert cwd == fake_repo.working_tree_dir
-
-    def test_triage_with_mock_claude_uses_execute_with_mock_command(self):
-        """When mock_claude is set, triage execute() receives a ClaudeCodeCommand with mock_command."""
-        triage_json = json.dumps({"will_fix": [], "needs_clarification": []})
-        triage_result = ClaudeResult(
-            returncode=0,
-            output=CapturedOutput(triage_json),
-            result_text=triage_json,
-        )
-
-        mock_path = "/path/to/mock-claude"
-        processor, _, fake_repo, _, fake_claude = _make_processor(
-            comments=[{"id": 1, "body": "fix this", "user": {"login": "u"}}],
-            mock_claude=mock_path,
-        )
-        fake_claude.set_result(triage_result)
-
-        processor.process_pr_feedback()
-
-        assert len(fake_claude.calls) == 1
-        method, cmd, cwd = fake_claude.calls[0]
-        assert method == "execute"
-        assert isinstance(cmd, ClaudeCodeCommand)
-        assert cmd.cwd == fake_repo.working_tree_dir
-        assert cmd.mock_command == [mock_path, "triage-42"]
-        assert cwd == fake_repo.working_tree_dir
-
-    def test_triage_reads_from_result_text(self):
-        """Triage parses will_fix from result.result_text, not output.stdout."""
-        triage_json = json.dumps({
-            "will_fix": [{"comment_ids": [1], "description": "fix"}],
-            "needs_clarification": [],
-        })
-        triage_result = ClaudeResult(
-            returncode=0,
-            output=CapturedOutput("ignored raw stdout"),
-            result_text=triage_json,
-        )
-        fix_result = ClaudeResult(returncode=0)
-
-        processor, _, fake_repo, _, fake_claude = _make_processor(
-            comments=[{"id": 1, "body": "fix this", "user": {"login": "u"}}],
-        )
-        fake_repo.set_head_sha("aaa111")
-
-        def advance_head():
-            fake_repo.set_head_sha("bbb222")
-
-        fake_claude.set_results([triage_result, fix_result])
-        fake_claude.set_side_effects([lambda: None, advance_head])
-
-        had_feedback, made_changes = processor.process_pr_feedback()
-        assert had_feedback is True
-        assert made_changes is True
-
-    def test_extract_result_text_no_longer_defined(self):
-        """_extract_result_text private function is removed from the processor."""
-        assert not hasattr(PullRequestReviewProcessor, "_extract_result_text")
-
-
-SINGLE_COMMENT = [{"id": 1, "body": "fix this", "user": {"login": "u"}}]
-
-
-@pytest.mark.unit
-class TestProcessPrFeedbackFixGroup:
-    """process_pr_feedback processes fix groups using injected collaborators."""
-
-    def test_fix_uses_git_repo_head_sha_not_gitrepo(self):
-        """HEAD tracking uses self._git_repo.head_sha, not GitRepo(worktree_path)."""
-        processor, _, _, _ = _make_fix_processor(SINGLE_COMMENT, _triage_with_fix([1]))
-        had_feedback, made_changes = processor.process_pr_feedback()
-        assert had_feedback is True
-        assert made_changes is True
-
-    def test_fix_uses_git_repo_push_not_push_branch_to_remote(self):
-        """Push uses self._git_repo.push(), not push_branch_to_remote()."""
-        processor, fake_repo, _, _ = _make_fix_processor(SINGLE_COMMENT, _triage_with_fix([1]))
-        processor.process_pr_feedback()
-        push_calls = [c for c in fake_repo.calls if c[0] == "push"]
-        assert len(push_calls) == 1
-
-    def test_fix_uses_execute_with_command_dataclass(self):
-        """Non-mock fix dispatches execute() with the CommandBuilder fix command."""
-        from i2code.implement.command_builder import FixRequest
-        processor, fake_repo, _, fake_claude = _make_fix_processor(
-            SINGLE_COMMENT, _triage_with_fix([1]),
-        )
-        processor.process_pr_feedback()
-
-        assert len(fake_claude.calls) == 2
-        triage_method, _, _ = fake_claude.calls[0]
-        fix_method, fix_cmd, fix_cwd = fake_claude.calls[1]
-        assert triage_method == "execute"
-        assert fix_method == "execute"
-        assert isinstance(fix_cmd, ClaudeCodeCommand)
-        feedback_content = PullRequestReviewProcessor._format_all_feedback(
-            SINGLE_COMMENT, [], [],
-        )
-        expected = CommandBuilder().build_fix_command(
-            FixRequest(
-                pr_url="https://github.com/org/repo/pull/42",
-                feedback_content=feedback_content,
-                fix_description="Fix issue",
-            ),
-            cwd=fake_repo.working_tree_dir,
-            interactive=False,
-        )
-        assert fix_cmd == expected
-        assert fix_cwd == fake_repo.working_tree_dir
-
-    def test_fix_with_mock_claude_uses_execute_with_mock_command(self):
-        """When mock_claude is set, fix execute() receives a ClaudeCodeCommand with mock_command."""
-        mock_path = "/path/to/mock-claude"
-        processor, fake_repo, _, fake_claude = _make_fix_processor(
-            SINGLE_COMMENT, _triage_with_fix([1]), mock_claude=mock_path,
-        )
-        processor.process_pr_feedback()
-
-        assert len(fake_claude.calls) == 2
-        fix_method, fix_cmd, fix_cwd = fake_claude.calls[1]
-        assert fix_method == "execute"
-        assert isinstance(fix_cmd, ClaudeCodeCommand)
-        assert fix_cmd.cwd == fake_repo.working_tree_dir
-        assert fix_cmd.mock_command == [mock_path, "fix-42-1"]
-        assert fix_cwd == fake_repo.working_tree_dir
-
-    def test_fix_marks_all_feedback_processed(self):
-        """After processing, all feedback IDs are marked processed in state."""
-        processor, _, fake_state, _ = _make_fix_processor(
-            comments=[{"id": 1, "body": "fix this", "user": {"login": "u"}}],
-            triage_json=_triage_with_fix([1]),
-            reviews=[{"id": 10, "body": "looks bad", "state": "CHANGES_REQUESTED", "user": {"login": "u"}}],
-            conversation_comments=[{"id": 20, "body": "general note", "user": {"login": "u"}}],
-        )
-        processor.process_pr_feedback()
-        assert 1 in fake_state.processed_comment_ids
-        assert 10 in fake_state.processed_review_ids
-        assert 20 in fake_state.processed_conversation_ids
-
-
-@pytest.mark.unit
-class TestFormatAllFeedback:
-    """Test formatting all feedback types for Claude."""
-
-    def test_format_all_feedback_includes_reviews(self):
-        """Should format PR reviews with state and body."""
-        reviews = [
-            {"id": 1, "state": "CHANGES_REQUESTED", "body": "Please fix the tests",
-             "user": {"login": "reviewer1"}}
-        ]
-        result = PullRequestReviewProcessor._format_all_feedback([], reviews, [])
-        assert "## PR Reviews" in result
-        assert "CHANGES_REQUESTED" in result
-        assert "Please fix the tests" in result
-        assert "reviewer1" in result
-
-    def test_format_all_feedback_includes_review_comments(self):
-        """Should format review comments with file path and line."""
-        review_comments = [
-            {"id": 2, "body": "This variable name is unclear",
-             "path": "src/main.py", "line": 42, "user": {"login": "reviewer2"}}
-        ]
-        result = PullRequestReviewProcessor._format_all_feedback(review_comments, [], [])
-        assert "## Review Comments" in result
-        assert "src/main.py:42" in result
-        assert "This variable name is unclear" in result
-
-    def test_format_all_feedback_includes_conversation_comments(self):
-        """Should format general PR comments."""
-        conversation_comments = [
-            {"id": 3, "body": "Great work overall!", "user": {"login": "lead"}}
-        ]
-        result = PullRequestReviewProcessor._format_all_feedback([], [], conversation_comments)
-        assert "## General PR Comments" in result
-        assert "Great work overall!" in result
-        assert "lead" in result
-
-    def test_format_all_feedback_combines_all_types(self):
-        """Should combine all feedback types into one formatted string."""
-        reviews = [{"id": 1, "state": "APPROVED", "body": "LGTM", "user": {"login": "r1"}}]
-        review_comments = [{"id": 2, "body": "Nitpick", "path": "a.py", "line": 1, "user": {"login": "r2"}}]
-        conversation_comments = [{"id": 3, "body": "Thanks", "user": {"login": "r3"}}]
-        result = PullRequestReviewProcessor._format_all_feedback(review_comments, reviews, conversation_comments)
-        assert "## PR Reviews" in result
-        assert "## Review Comments" in result
-        assert "## General PR Comments" in result
-
-
-@pytest.mark.unit
-class TestGetNewFeedback:
-    """Test filtering feedback to unprocessed items."""
-
-    def test_get_new_feedback_filters_processed(self):
-        """Should filter out already processed feedback."""
-        all_feedback = [{"id": 1, "body": "Old"}, {"id": 2, "body": "New"}, {"id": 3, "body": "Another"}]
-        new_feedback = PullRequestReviewProcessor._get_new_feedback(all_feedback, [1])
-        assert len(new_feedback) == 2
-        assert all(f["id"] in [2, 3] for f in new_feedback)
-
-    def test_get_new_feedback_returns_all_when_none_processed(self):
-        """Should return all feedback when nothing processed yet."""
-        all_feedback = [{"id": 1, "body": "Comment 1"}, {"id": 2, "body": "Comment 2"}]
-        assert len(PullRequestReviewProcessor._get_new_feedback(all_feedback, [])) == 2
-
-    def test_get_new_feedback_returns_empty_when_all_processed(self):
-        """Should return empty list when all feedback processed."""
-        assert PullRequestReviewProcessor._get_new_feedback([{"id": 1, "body": "Comment"}], [1]) == []
-
-
-@pytest.mark.unit
-class TestParseOwnerRepo:
-    """Test parsing owner/repo from git remote URLs."""
-
-    def test_https_with_credentials(self):
-        assert parse_owner_repo(
-            "https://x-access-token:ghs_abc123@github.com/my-org/my-repo.git"
-        ) == ("my-org", "my-repo")
-
-    def test_ssh(self):
-        assert parse_owner_repo(
-            "git@github.com:my-org/my-repo.git"
-        ) == ("my-org", "my-repo")
-
-    def test_https(self):
-        assert parse_owner_repo(
-            "https://github.com/my-org/my-repo.git"
-        ) == ("my-org", "my-repo")
-
-    def test_https_without_dot_git(self):
-        assert parse_owner_repo(
-            "https://github.com/my-org/my-repo"
-        ) == ("my-org", "my-repo")
-
-
-@pytest.mark.unit
-class TestParseTriageResult:
-    """Test parsing JSON triage result from Claude."""
-
-    def test_parse_triage_result_with_json_code_block(self):
-        """Should parse JSON from markdown code block."""
-        output = '''Here's the triage:
-```json
-{
-  "will_fix": [{"comment_ids": [1, 2], "description": "Fix typo"}],
-  "needs_clarification": []
-}
-```
-'''
-        result = PullRequestReviewProcessor._parse_triage_result(output)
-        assert result is not None
-        assert len(result["will_fix"]) == 1
-        assert result["will_fix"][0]["comment_ids"] == [1, 2]
-
-    def test_parse_triage_result_with_plain_json(self):
-        """Should parse plain JSON output."""
-        output = '{"will_fix": [], "needs_clarification": [{"comment_id": 5, "question": "What?"}]}'
-        result = PullRequestReviewProcessor._parse_triage_result(output)
-        assert result is not None
-        assert len(result["needs_clarification"]) == 1
-        assert result["needs_clarification"][0]["comment_id"] == 5
-
-    def test_parse_triage_result_returns_none_on_invalid(self):
-        """Should return None for invalid JSON."""
-        assert PullRequestReviewProcessor._parse_triage_result("This is not JSON at all") is None
-
-
-@pytest.mark.unit
-class TestGetFeedbackByIds:
-    """Test filtering feedback by IDs."""
-
-    def test_get_feedback_by_ids_returns_matching(self):
-        """Should return only feedback with matching IDs."""
-        all_feedback = [{"id": 1, "body": "C1"}, {"id": 2, "body": "C2"}, {"id": 3, "body": "C3"}]
-        result = PullRequestReviewProcessor._get_feedback_by_ids(all_feedback, [1, 3])
-        assert len(result) == 2
-        assert result[0]["id"] == 1
-        assert result[1]["id"] == 3
-
-    def test_get_feedback_by_ids_returns_empty_for_no_matches(self):
-        """Should return empty list when no IDs match."""
-        assert PullRequestReviewProcessor._get_feedback_by_ids([{"id": 1, "body": "Comment"}], [99]) == []
-
-
-@pytest.mark.unit
-class TestDetermineCommentType:
-    """Test determining comment type from ID."""
-
-    def test_determine_comment_type_review(self):
-        """Should return 'review' for review comment IDs."""
-        result = PullRequestReviewProcessor._determine_comment_type(
-            100, [{"id": 100, "body": "Review"}], [{"id": 200, "body": "General"}],
-        )
-        assert result == "review"
-
-    def test_determine_comment_type_conversation(self):
-        """Should return 'conversation' for non-review comment IDs."""
-        result = PullRequestReviewProcessor._determine_comment_type(
-            200, [{"id": 100, "body": "Review"}], [{"id": 200, "body": "General"}],
-        )
-        assert result == "conversation"
-
-
-I2CODE_MARKER = "<!-- i2code -->"
-
-
-def _get_reply_body(fake_gh, call_name):
-    """Extract the reply body from the single matching call on fake_gh."""
-    reply_calls = [c for c in fake_gh.calls if c[0] == call_name]
-    assert len(reply_calls) == 1, f"Expected 1 {call_name} call, got {len(reply_calls)}"
-    body_index = 3 if call_name == "reply_to_review_comment" else 2
-    return reply_calls[0][body_index]
-
-
-def _assert_marker_prefixed(body, *expected_substrings):
-    """Assert body starts with the i2code marker and contains expected substrings."""
-    assert body.startswith(f"{I2CODE_MARKER}\n"), f"Expected marker prefix, got: {body!r}"
-    for substring in expected_substrings:
-        assert substring in body, f"Expected {substring!r} in body: {body!r}"
-
-
-def _push_and_reply_body(comment_id, review_comments, conversation_comments):
-    """Call _push_and_reply and return the reply body posted to the matching call."""
-    processor, fake_gh, fake_repo, _, _ = _make_processor()
-    fake_repo.set_head_sha("aaa111")
-    processor._push_and_reply("abc12345", [comment_id], review_comments, conversation_comments)
-    call_name = "reply_to_review_comment" if review_comments else "reply_to_pr_comment"
-    return _get_reply_body(fake_gh, call_name)
-
-
-def _clarification_reply_body(comment_id, question, review_comments, conversation_comments):
-    """Call _reply_with_clarifications and return the reply body."""
-    processor, fake_gh, _, _, _ = _make_processor()
-    processor._reply_with_clarifications(
-        [{"comment_id": comment_id, "question": question}],
-        42, review_comments, conversation_comments,
-    )
-    call_name = "reply_to_review_comment" if review_comments else "reply_to_pr_comment"
-    return _get_reply_body(fake_gh, call_name)
-
-
-@pytest.mark.unit
-class TestPushAndReplyMarker:
-    """_push_and_reply prepends i2code marker to reply bodies."""
-
-    def test_review_comment_reply_starts_with_marker(self):
-        body = _push_and_reply_body(100, [{"id": 100, "body": "Fix this"}], [])
-        _assert_marker_prefixed(body, "Fixed in abc12345")
-
-    def test_conversation_comment_reply_starts_with_marker(self):
-        body = _push_and_reply_body(200, [], [{"id": 200, "body": "General note"}])
-        _assert_marker_prefixed(body)
-
-
-@pytest.mark.unit
-class TestReplyWithClarificationsMarker:
-    """_reply_with_clarifications prepends i2code marker to clarification replies."""
-
-    def test_review_comment_clarification_starts_with_marker(self):
-        body = _clarification_reply_body(100, "Could you elaborate?", [{"id": 100, "body": "What does this do?"}], [])
-        _assert_marker_prefixed(body, "Could you elaborate?")
-
-    def test_conversation_comment_clarification_starts_with_marker(self):
-        body = _clarification_reply_body(200, "Can you explain further?", [], [{"id": 200, "body": "Why this approach?"}])
-        _assert_marker_prefixed(body, "Re: comment 200", "Can you explain further?")
-
-
-@pytest.mark.unit
-class TestSelfCommentFiltering:
-    """_fetch_unprocessed_feedback filters out comments with i2code marker."""
-
-    def test_filters_out_marker_comments_and_marks_processed(self):
-        """Comments starting with <!-- i2code --> are excluded and their IDs marked processed."""
-        user_comment = {"id": 1, "body": "Please fix the typo", "user": {"login": "reviewer"}}
-        self_comment = {"id": 2, "body": "<!-- i2code -->\nFixed in abc123", "user": {"login": "bot"}}
-        another_user_comment = {"id": 3, "body": "Also check line 10", "user": {"login": "reviewer"}}
-
-        processor, _, _, fake_state, _ = _make_processor(
-            comments=[user_comment, self_comment, another_user_comment],
-            reviews=[],
-            conversation_comments=[],
-        )
-
-        review_comments, reviews, conversation = processor._fetch_unprocessed_feedback(42)
-
-        assert len(review_comments) == 2
-        assert all(c["id"] in [1, 3] for c in review_comments)
-        assert 2 in fake_state.processed_comment_ids
-
-    def test_filters_marker_from_conversation_comments(self):
-        """Marker filtering also applies to conversation comments."""
-        user_comment = {"id": 10, "body": "Nice work", "user": {"login": "reviewer"}}
-        self_comment = {"id": 11, "body": "<!-- i2code -->\nRe: comment 10\n\nClarification", "user": {"login": "bot"}}
-
-        processor, _, _, fake_state, _ = _make_processor(
-            comments=[],
-            reviews=[],
-            conversation_comments=[user_comment, self_comment],
-        )
-
-        review_comments, reviews, conversation = processor._fetch_unprocessed_feedback(42)
-
-        assert len(conversation) == 1
-        assert conversation[0]["id"] == 10
-        assert 11 in fake_state.processed_conversation_ids
-
-    def test_user_followup_after_i2code_reply_is_included(self):
-        """Scenario 4: user comment → i2code reply → user follow-up.
-
-        The i2code reply is filtered out but both user comments are returned.
-        """
-        user_comment = {"id": 1, "body": "Why is this function public?", "user": {"login": "reviewer"}}
-        i2code_reply = {"id": 2, "body": "<!-- i2code -->\nIt's public because it's called from integration tests.", "user": {"login": "reviewer"}}
-        user_followup = {"id": 3, "body": "OK, but can we make it package-private instead?", "user": {"login": "reviewer"}}
-
-        processor, _, _, fake_state, _ = _make_processor(
-            comments=[user_comment, i2code_reply, user_followup],
-            reviews=[],
-            conversation_comments=[],
-        )
-
-        review_comments, reviews, conversation = processor._fetch_unprocessed_feedback(42)
-
-        assert len(review_comments) == 2
-        assert review_comments[0]["id"] == 1
-        assert review_comments[1]["id"] == 3
-        assert 2 in fake_state.processed_comment_ids
-        assert 1 not in fake_state.processed_comment_ids
-        assert 3 not in fake_state.processed_comment_ids
-
-
-@pytest.mark.unit
-class TestResolvedThreadFiltering:
-    """_fetch_unprocessed_feedback filters out comments in resolved review threads."""
-
-    def test_resolved_thread_comments_excluded_and_marked_processed(self):
-        """Comments whose IDs are in the resolved-thread set are excluded and marked processed."""
-        user_comment = {"id": 1, "body": "Please fix the typo", "user": {"login": "reviewer"}}
-        resolved_comment = {"id": 2, "body": "Consider a constant here", "user": {"login": "reviewer"}}
-        another_user_comment = {"id": 3, "body": "Also check line 10", "user": {"login": "reviewer"}}
-
-        processor, fake_gh, _, fake_state, _ = _make_processor(
-            comments=[user_comment, resolved_comment, another_user_comment],
-            reviews=[],
-            conversation_comments=[],
-        )
-        fake_gh.set_resolved_review_comment_ids("test", "repo", 42, {2})
-
-        review_comments, reviews, conversation = processor._fetch_unprocessed_feedback(42)
-
-        assert len(review_comments) == 2
-        assert all(c["id"] in [1, 3] for c in review_comments)
-        assert 2 in fake_state.processed_comment_ids
-        assert 1 not in fake_state.processed_comment_ids
-        assert 3 not in fake_state.processed_comment_ids
-
-    def test_mixed_resolved_and_unresolved_threads_returns_only_unresolved(self):
-        """Scenario 6: three threads (A resolved, B unresolved, C resolved) — only B returned."""
-        thread_a_comment = {"id": 10, "body": "Thread A feedback", "user": {"login": "reviewer"}}
-        thread_b_comment = {"id": 20, "body": "Thread B feedback", "user": {"login": "reviewer"}}
-        thread_c_comment = {"id": 30, "body": "Thread C feedback", "user": {"login": "reviewer"}}
-
-        processor, fake_gh, _, fake_state, _ = _make_processor(
-            comments=[thread_a_comment, thread_b_comment, thread_c_comment],
-            reviews=[],
-            conversation_comments=[],
-        )
-        fake_gh.set_resolved_review_comment_ids("test", "repo", 42, {10, 30})
-
-        review_comments, reviews, conversation = processor._fetch_unprocessed_feedback(42)
-
-        assert len(review_comments) == 1
-        assert review_comments[0]["id"] == 20
-        assert 10 in fake_state.processed_comment_ids
-        assert 30 in fake_state.processed_comment_ids
-        assert 20 not in fake_state.processed_comment_ids
-
-
-PR6_REPO = "humansintheloop-dev/humansintheloop-dev-workflow-and-tools"
-PR6_NUMBER = 6
-
-
-@pytest.mark.unit
-class TestTriageFeedbackLogging:
-    """Test that _triage_feedback logs prompt and response to file."""
-
-    @pytest.fixture
-    def pr6_feedback(self):
-        assert os.path.exists(PR6_FEEDBACK_FILE), (
-            f"Fixture not found: {PR6_FEEDBACK_FILE}\n"
-            f"Run: uv run --with pytest pytest tests/implement/test_triage_failure_logging.py -m gather_fixtures"
-        )
-        with open(PR6_FEEDBACK_FILE) as f:
-            return json.load(f)
-
-    @pytest.fixture
-    def home_dir(self, tmp_path, monkeypatch):
-        monkeypatch.setattr("pathlib.Path.home", classmethod(lambda cls: tmp_path))
-        return tmp_path
-
-    def _make_triage_processor(self, claude_stdout):
-        worktree_name = "myrepo-wt-improve-modularity"
-
-        fake_gh = FakeGitHubClient()
-        fake_gh.set_pr_url(PR6_NUMBER, f"https://github.com/{PR6_REPO}/pull/{PR6_NUMBER}")
-
-        fake_repo = FakeGitRepository(
-            working_tree_dir=f"/worktrees/{worktree_name}", gh_client=fake_gh,
-        )
-        fake_repo.pr_number = PR6_NUMBER
-        fake_repo.branch = "idea/improve-modularity/01-extract-ideaproject-class"
-        fake_repo.set_pushed(True)
-
-        fake_claude = FakeClaudeRunner()
-        _diagnostics, result_text = _parse_stream_json_output(claude_stdout)
-        fake_claude.set_result(ClaudeResult(
-            returncode=0,
-            output=CapturedOutput(claude_stdout),
-            result_text=result_text,
-        ))
-
-        processor = PullRequestReviewProcessor(
-            opts=ImplementOpts(idea_directory="/tmp/idea", non_interactive=True, skip_ci_wait=True),
-            git_repo=fake_repo, state=FakeWorkflowState(), claude_runner=fake_claude,
-        )
-
-        return processor, worktree_name
-
-    def test_logs_feedback_content_on_triage_parse_failure(self, pr6_feedback, home_dir):
-        processor, worktree_name = self._make_triage_processor("not json at all")
-
-        review_comments = pr6_feedback["review_comments"]
-        feedback_content = PullRequestReviewProcessor._format_all_feedback(
-            review_comments, pr6_feedback["reviews"], pr6_feedback["conversation_comments"],
-        )
-
-        result = processor._triage_feedback(feedback_content, PR6_NUMBER)
-        assert result is None
-
-        log_file = home_dir / ".hitl" / worktree_name / "logs" / "log.log"
-        assert log_file.exists(), f"Expected log file at {log_file}"
-
-        log_content = log_file.read_text()
-        assert "--- prompt ---" in log_content
-        assert "--- claude response ---" in log_content
-        assert "not json at all" in log_content
-
-        for comment in review_comments:
-            body = comment.get("body", "").strip()
-            if body:
-                assert body in log_content, f"Expected review comment body in log: {body[:60]}..."
-                break
-
-    def test_parses_real_claude_triage_response(self, pr6_feedback, home_dir):
-        with open(PR6_TRIAGE_STDOUT_FILE) as f:
-            triage_stdout = f.read()
-
-        processor, _ = self._make_triage_processor(triage_stdout)
-
-        feedback_content = PullRequestReviewProcessor._format_all_feedback(
-            pr6_feedback["review_comments"], pr6_feedback["reviews"], pr6_feedback["conversation_comments"],
-        )
-
-        result = processor._triage_feedback(feedback_content, PR6_NUMBER)
-        assert result is not None, "Expected triage to parse successfully"
-        assert "will_fix" in result
-        assert "needs_clarification" in result
-        assert len(result["will_fix"]) > 0
